Fix remaining CPU operators for non-standard bools
In addition to using `c10::load` in more kernels, this specializes
`Vectorized<bool>::loadu` to first load as `int8_t` then convert to
bool, in the same way as `c10::load` does.

For `unique_cpu`, the values are loaded inside of the
`std::unordered_set` constuctor, so I've added a separate function
specifically for bools. This fixes the loading issue and is faster
because it uses the fact that bools can only have one of two values.

ghstack-source-id: 44004bb777d192b0df718d78d05559260c86d36c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/79390