[ci] don't mutate BUILD_ENVIRONMENT in test.sh
We append TEST_CONFIG to BUILD_ENVIRONMENT at the beginning of test.sh

This is bad because it alters a fundamental env var, and also isn't
really necessary because it doesn't add information.

In most case where we check BUILD_ENV, we also check TEST_CONFIG, making
it redundant anyway.

ghstack-source-id: 56348abd7f8eff759f349d6cdf43d04ba6e5605b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/80144