[ci] delete JOB_BASE_NAME
`JOB_BASE_NAME` was a holdover from jenkins compatibility. Eventually,
it morphed to be always set to the build enviroment + `-test` or
`-build`, and we used it to detect whether we were in a build or test.

That's sort of pointless, so removing and fixing up the few remaining
use cases.

ghstack-source-id: d8d48a4ab4cfcaf3c070962fc6ea2062155c814d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/80046