Make linalg.inv composite of linalg.solve
The `getri` kernel calls inside `getrs` so we can do so explicitly
ourselves and save ourselves from having to maintain an extra kernel.
This way we just need to optimise `lu_factor` and `lu_solve` and `inv`
will be as efficient as it can be, as it'll be choosing the best backend
to perform the factorisation and the best backend (not necessarily the
same) to perform the solve.

ghstack-source-id: 83015fbfed5e37394c1b8245bb5dc0a73b904b2e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/80074